CLI: extract followJsonlStream + ago helpers (#101)
Mirrors the cross-package reuse pass on the plugin side:

cli/src/format/stream.mjs — followJsonlStream(url, onLine)
factors out the SIGINT-handling + getStreamLines + try/catch boil
that followTestStream and followCoverageStream both copy-pasted.
Caller decides --json passthrough and per-line state (test stream
still tracks hasFailed for the exit code, coverage stream is pure
formatter).

cli/src/format/relative-time.mjs — ago(ms) was duplicated in
format/coverage-state.mjs and commands/test-sessions.mjs. Single
source now; test-sessions used to round-truncate at hours, this
adds the days bucket as a side effect (matches coverage's version,
which was the more correct one for old runs).

Tests: 779/779 CLI.
