+++ cli/src/commands/test-sessions.mjs
@@ -2,6 +2,7 @@ import { get } from "../client.mjs";
 import { output } from "../output.mjs";
 import { green, red, yellow } from "../color.mjs";
 import { formatTable } from "../format/table.mjs";
+import { ago } from "../format/relative-time.mjs";
 
 /**
  * List test runs (active and completed).
@@ -41,15 +42,6 @@ function formatCounts(s) {
   if (s.errors > 0) parts.push(red(`${s.errors} errors`));
   if (s.ignored > 0) parts.push(yellow(`${s.ignored} ign`));
   return parts.join(", ");
-}
-
-function ago(ms) {
-  const s = Math.floor(ms / 1000);
-  if (s < 60) return `${s}s ago`;
-  const m = Math.floor(s / 60);
-  if (m < 60) return `${m}m ago`;
-  const h = Math.floor(m / 60);
-  return `${h}h ago`;
 }
 
 export const help = `List test runs (active and completed).
