+++ cli/src/format/coverage-state.mjs
@@ -5,6 +5,7 @@
 //   A — exactly one launch/origin token
 //   B — zero or more data/analysis tokens, in order
 //   C — at most one relative-time token (live runs only)
+import { ago } from "./relative-time.mjs";
 
 /**
  * @param {object} run — entry from /coverage/runs (one element)
@@ -64,14 +65,3 @@ export function composeStatus(run, now = Date.now()) {
   return tokens.join(", ");
 }
 
-function ago(ms) {
-  if (ms < 0) ms = 0;
-  const s = Math.floor(ms / 1000);
-  if (s < 60) return `${s}s ago`;
-  const m = Math.floor(s / 60);
-  if (m < 60) return `${m}m ago`;
-  const h = Math.floor(m / 60);
-  if (h < 24) return `${h}h ago`;
-  const d = Math.floor(h / 24);
-  return `${d}d ago`;
-}
