+++ cli/src/format/relative-time.mjs
@@ -1,0 +1,13 @@
+/** Render a millisecond duration as a coarse-grained "Xs/Xm/Xh/Xd ago"
+ *  string. Used by table renderers that surface relative timing. */
+export function ago(ms) {
+  if (ms < 0) ms = 0;
+  const s = Math.floor(ms / 1000);
+  if (s < 60) return `${s}s ago`;
+  const m = Math.floor(s / 60);
+  if (m < 60) return `${m}m ago`;
+  const h = Math.floor(m / 60);
+  if (h < 24) return `${h}h ago`;
+  const d = Math.floor(h / 24);
+  return `${d}d ago`;
+}
